Moved default values to model
