+++ src/main/kotlin/ui/views/SettingsPane.kt
@@ -3,19 +3,15 @@ package ui.views
 import javafx.scene.paint.Color
 import tornadofx.*
 import ui.controllers.ConversationController
-import ui.viewmodels.SettingsViewModel
+import ui.viewmodels.SettingsModel
 import ui.viewmodels.validateDB
-import java.io.File
 
 class SettingsPane : View("Settings") {
 
-    private val model: SettingsViewModel by inject()
+    private val model: SettingsModel by inject()
     private val controller: ConversationController by inject()
 
     init {
-        model.messageDB.value = "${System.getProperty("user.home")}/Library/Messages/chat.db"
-        model.addressBookDB.value = findAddressBook()
-
         runLater {
             model.validationContext.validate()
         }
@@ -56,16 +52,4 @@ class SettingsPane : View("Settings") {
             }
         }
     }
-}
-
-fun findAddressBook(): String? {
-    val rootDir = "${System.getProperty("user.home")}/Library/Application Support/AddressBook/Sources"
-
-    //For now we'll assume the biggest address book is the one we want.
-    //In the future we'd like to merge all available address books
-    val db = File(rootDir).walk()
-        .filter { it.toString().endsWith("db") }
-        .maxBy { it.length() }
-
-    return db?.path
 }
